api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
supabase==2.3.4
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.2
pydantic==2.5.3
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
